Issue 253: Create XmlXPathRetrieve to retrieve XML data using XPath expressions
- Fixed review items.
- Modified comment for class to reflect value returned if no xml data
matching the expression
- Renamed XPathRetrieve to XmlXPathRetrieve
- removed XmlXPathElement and moved XmlXPathRetrieve to xml package
